requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.6.0
python-multipart>=0.0.9
google-cloud-storage>=2.14.0
mplfinance>=0.12.10b0
matplotlib>=3.8.0